🎨 リファクタリング
