+++ package.json
@@ -12,6 +12,7 @@
     "ejs": "^3.1.10",
     "ejs-mate": "^4.0.0",
     "express": "^4.19.2",
+    "express-session": "^1.18.0",
     "joi": "^17.13.3",
     "method-override": "^3.0.0",
     "mongoose": "^5.13.22"
