+++ routes/campgrounds.js
@@ -1,0 +1,77 @@
+const express = require("express");
+const router = express.Router();
+const catchAsync = require("../utils/catchAsync");
+const ExpressError = require("../utils/ExpressError");
+const Campground = require("../models/campground");
+const { campgroundSchema } = require("../schemas");
+
+const validateCampground = (req, res, next) => {
+  const { error } = campgroundSchema.validate(req.body);
+  if (error) {
+    const msg = error.details.map((detail) => detail.message).join(",");
+    throw new ExpressError(msg, 400);
+  } else {
+    next();
+  }
+};
+
+router.get("/", async (req, res) => {
+  const campgrounds = await Campground.find({});
+  res.render("campgrounds/index", { campgrounds });
+});
+
+router.get("/new", (req, res) => {
+  res.render("campgrounds/new");
+});
+
+router.get(
+  "/:id",
+  catchAsync(async (req, res) => {
+    const campground = await Campground.findById(req.params.id).populate(
+      "reviews",
+    );
+    res.render("campgrounds/show", { campground });
+  }),
+);
+
+router.post(
+  "/",
+  validateCampground,
+  catchAsync(async (req, res) => {
+    // if (!req.body.campground) throw new ExpressError("不正なキャンプ場のデータです", 400)
+    const campground = new Campground(req.body.campground);
+    await campground.save();
+    res.redirect(`/campgrounds/${campground._id}`);
+  }),
+);
+
+router.get(
+  "/:id/edit",
+  catchAsync(async (req, res) => {
+    const campground = await Campground.findById(req.params.id);
+    res.render("campgrounds/edit", { campground });
+  }),
+);
+
+router.put(
+  "/:id",
+  validateCampground,
+  catchAsync(async (req, res) => {
+    const { id } = req.params;
+    const campground = await Campground.findByIdAndUpdate(id, {
+      ...req.body.campground,
+    });
+    res.redirect(`/campgrounds/${campground._id}`);
+  }),
+);
+
+router.delete(
+  "/:id",
+  catchAsync(async (req, res) => {
+    const { id } = req.params;
+    await Campground.findByIdAndDelete(id);
+    res.redirect("/campgrounds");
+  }),
+);
+
+module.exports = router;
